Change preferred pattern names
* standalone -> instance per cluster
* namespaced -> instance per namespace

Because: On top of that there is the option of using one instance that handles several namespaces/tenants (e.g. "Applications in any namespace" feature in Argo CD). This is not an explicit pattern, though, because it uses one instance per cluster. But it collides with "standalone" and "namespaced". The new names remove this confusion. Hopefully.